+++ cmd/hatchet-engine/engine/run.go
@@ -126,7 +126,7 @@ func RunWithConfig(ctx context.Context, sc *server.ServerConfig, cleanup *cleanu
 func runV0Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.Cleanup) error {
 	var l = sc.Logger
 
-	shutdown, err := telemetry.InitTracer(&telemetry.TracerOpts{
+	telemetryShutdown, err := telemetry.InitTracer(&telemetry.TracerOpts{
 		ServiceName:   sc.OpenTelemetry.ServiceName,
 		CollectorURL:  sc.OpenTelemetry.CollectorURL,
 		TraceIdRatio:  sc.OpenTelemetry.TraceIdRatio,
@@ -510,9 +510,7 @@ func runV0Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 	}
 
 	cleanup.Add(
-		func() error {
-			return shutdown(ctx)
-		},
+		telemetryShutdown,
 		"telemetry",
 	)
 
@@ -530,7 +528,7 @@ func runV0Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 func runV1Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.Cleanup) error {
 	var l = sc.Logger
 
-	shutdown, err := telemetry.InitTracer(&telemetry.TracerOpts{
+	telemetryShutdown, err := telemetry.InitTracer(&telemetry.TracerOpts{
 		ServiceName:   sc.OpenTelemetry.ServiceName,
 		CollectorURL:  sc.OpenTelemetry.CollectorURL,
 		TraceIdRatio:  sc.OpenTelemetry.TraceIdRatio,
@@ -943,9 +941,7 @@ func runV1Config(ctx context.Context, sc *server.ServerConfig, cleanup *cleanup.
 	}
 
 	cleanup.Add(
-		func() error {
-			return shutdown(ctx)
-		},
+		telemetryShutdown,
 		"telemetry",
 	)