+++ pkg/config/loader/loader.go
@@ -559,6 +559,7 @@ func createControllerLayer(dc *database.Layer, cf *server.ServerConfigFile, vers
 		tenantId := tenant.ID
 
 		analyticsEmitter.Tenant(tenantId, map[string]interface{}{
+			"id":   tenantId.String(),
 			"name": tenant.Name,
 			"slug": tenant.Slug,
 		})
@@ -689,6 +690,7 @@ func createControllerLayer(dc *database.Layer, cf *server.ServerConfigFile, vers
 		cf.Runtime.SchedulerConcurrencyRateLimit,
 		cf.Runtime.SchedulerConcurrencyPollingMinInterval,
 		cf.Runtime.SchedulerConcurrencyPollingMaxInterval,
+		cf.Runtime.SchedulerAdvisoryLockTimeout,
 		cf.Runtime.OptimisticSchedulingEnabled,
 		cf.Runtime.OptimisticSchedulingSlots,
 	)